Extract Sections of Same "Type" from Parse Tree

Extract Sections of Same "Type" from Parse Tree

get_elements_by_type(x, result = NULL)

Arguments

x
parse tree as returned by parse

result
optional. Result as returned by analyse

Examples

# Parse an R script file (here, a file from kwb.utils)
x <- parse("https://raw.githubusercontent.com/KWB-R/kwb.utils/master/R/column.R")

# For each "type" of code segment, extract all occurrences
elements <- get_elements_by_type(x)
#> Analysing the parse tree ... ok. (9.68s) 
# Show all for-loops
elements$`language|call|for|4|`
#> [[1]]
#> for (column in columns) {
#>     catAndRun(sprintf("Removing unused factors from %s$%s", data_name, 
#>         column), data[[column]] <- droplevels(data[[column]]), 
#>         dbg = dbg)
#> }
#> 
#> [[2]]
#> for (col in colNames) {
#>     if (!col %in% names(dataFrame)) {
#>         dataFrame[[col]] <- rep(fill.value, n)
#>     }
#> }
#> 
#> [[3]]
#> for (column in names(renamings)) {
#>     columns[columns == column] <- renamings[[column]]
#> }
#> 
#> [[4]]
#> for (column in columns) {
#>     n_digits <- if (is.list(digits)) {
#>         digits[[column]]
#>     }
#>     else {
#>         digits
#>     }
#>     dframe[[column]] <- round(dframe[[column]], digits = n_digits)
#> }
#> 
#> [[5]]
#> for (columnName in names(assignments)) {
#>     catIf(dbg, sprintf("Provide column '%s' to data frame '%s'... ", 
#>         columnName, name.x))
#>     .x[[columnName]] <- assignments[[columnName]]
#>     catIf(dbg, "ok.\n")
#> }
#> 

# Show all if-statements
elements$`language|call|if|3|`
#> [[1]]
#> if (!isNullOrEmpty(missing)) {
#>     plural_s_or_empty <- function(x) ifelse(length(x) > 1, "s", 
#>         "")
#>     sorted_list <- function(x) stringList(sort(x), collapse = "\n  ")
#>     infotext_missing <- sprintf("%d column%s missing in data frame '%s':\n  %s", 
#>         length(missing), plural_s_or_empty(missing), dataFrameName, 
#>         sorted_list(missing))
#>     infotext_available <- sprintf("Available column%s:\n  %s", 
#>         plural_s_or_empty(columnNames), sorted_list(columnNames))
#>     infotext <- paste(infotext_missing, infotext_available, sep = "\n")
#>     do.call(ifelse(do.stop, "stop", "warning"), list(infotext, 
#>         call. = FALSE))
#> }
#> 
#> [[2]]
#> if (!is.data.frame(data)) {
#>     stop("The object given to dropUnusedFactorLevels() is not a data frame but is", 
#>         "of\nclass: ", stringList(class(data)), call. = FALSE)
#> }
#> 
#> [[3]]
#> if (!col %in% names(dataFrame)) {
#>     dataFrame[[col]] <- rep(fill.value, n)
#> }
#> 
#> [[4]]
#> if (sum(c(before, after) != "") != 1) {
#>     stop("Exactly one of before and after must be given!")
#> }
#> 
#> [[5]]
#> if (is.null(columnNames) || length(columnNames[columnNames != 
#>     ""]) != length(newColumns)) {
#>     stop("All new columns must be named, i.e. given in the form 'name = values'")
#> }
#> 
#> [[6]]
#> if (!all(equalLength)) {
#>     stop("All new columns must have as many elements as Data has rows. ", 
#>         "This is not the case for: ", commaCollapsed(columnNames[!equalLength]))
#> }
#> 
#> [[7]]
#> if (isNullOrEmpty(tcol)) {
#>     warning("No POSIXt-column in data frame.")
#> }
#> 
#> [[8]]
#> if (is.null(columns) && !is.null(columnsToRemove)) {
#>     warning("The argument 'columnsToRemove' is deprecated. Please use the new ", 
#>         "argument 'columns' instead.", call. = FALSE)
#>     columns <- columnsToRemove
#> }
#> 
#> [[9]]
#> if (is.null(columns)) {
#>     if (is.null(pattern)) {
#>         stop("Either 'columns' or 'pattern' must be given to removeColumns()", 
#>             call. = FALSE)
#>     }
#>     else {
#>         columns <- grep(pattern, all_columns, value = TRUE)
#>     }
#> }
#> 
#> [[10]]
#> if (dbg) {
#>     columns_remove <- intersect(all_columns, columns)
#>     cat(sprintf("Removing column%s %s from '%s'.\n", ifelse(length(columns_remove) > 
#>         1, "s", ""), stringList(columns_remove), deparse(substitute(dframe))))
#> }
#> 
#> [[11]]
#> if (is.null(renamings)) {
#>     return(x)
#> }
#> 
#> [[12]]
#> if (!is.null(columnNames)) {
#>     warning("roundColumns: Argument 'columnNames' is deprecated. Use argument ", 
#>         "'columns' instead.", call. = FALSE)
#>     columns <- unique(c(columns, columnNames))
#> }
#> 
#> [[13]]
#> if (!is.null(columns)) {
#>     stopifnot(length(digits) == 1)
#> }
#> 
#> [[14]]
#> if (is.null(digits)) {
#>     return(dframe)
#> }
#> 
#> [[15]]
#> if (is.null(columns)) {
#>     columns <- names(which(sapply(dframe, is.numeric)))
#>     if (!is.null(pattern)) {
#>         columns <- grep(pattern, columns, value = TRUE)
#>     }
#> }
#> 
#> [[16]]
#> if (!is.null(pattern)) {
#>     columns <- grep(pattern, columns, value = TRUE)
#> }
#> 
#> [[17]]
#> if (!is.data.frame(x)) {
#>     stop(deparse(substitute(x)), " given to selectColumns() must be a data frame ", 
#>         "but is of class: ", stringList(class(x)), call. = FALSE)
#> }
#> 
#> [[18]]
#> if (is.null(columns) || length(columns) == 0 || all(is.na(columns))) {
#>     if (is.null(pattern)) {
#>         return(x)
#>     }
#>     else {
#>         columns <- grep(pattern, names(x), value = TRUE)
#>     }
#> }
#> 
#> [[19]]
#> if (!ok) {
#>     warning("Only the existing columns are selected.")
#>     columns <- intersect(columns, names(x))
#> }
#> 
#> [[20]]
#> if (any(is.unnamed(assignments))) {
#>     stop("All column assignments must be named!", call. = FALSE)
#> }
#> 

# Show all if-else-statements
elements$`language|call|if|4|`
#> [[1]]
#> if (length(columns)) {
#>     for (column in columns) {
#>         catAndRun(sprintf("Removing unused factors from %s$%s", 
#>             data_name, column), data[[column]] <- droplevels(data[[column]]), 
#>             dbg = dbg)
#>     }
#> } else {
#>     catIf(dbg, "No factor columns contained in data frame '", 
#>         data_name, "'")
#> }
#> 
#> [[2]]
#> if (before != "") {
#>     part1 <- if (i == 1) {
#>         partBetween
#>     }
#>     else {
#>         cbind(Data[, 1:(i - 1), drop = FALSE], partBetween)
#>     }
#>     part2 <- Data[, i:n.col, drop = FALSE]
#> } else {
#>     part1 <- Data[, 1:i, drop = FALSE]
#>     part2 <- if (i == n.col) {
#>         partBetween
#>     }
#>     else {
#>         cbind(partBetween, Data[, (i + 1):n.col, drop = FALSE], 
#>             stringsAsFactors = stringsAsFactors)
#>     }
#> }
#> 
#> [[3]]
#> if (i == 1) {
#>     partBetween
#> } else {
#>     cbind(Data[, 1:(i - 1), drop = FALSE], partBetween)
#> }
#> 
#> [[4]]
#> if (i == n.col) {
#>     partBetween
#> } else {
#>     cbind(partBetween, Data[, (i + 1):n.col, drop = FALSE], stringsAsFactors = stringsAsFactors)
#> }
#> 
#> [[5]]
#> if (length(columns) > 1) {
#>     args <- selectColumns(x, columns, ...)
#>     do.call(paste, c(args, sep = sep))
#> } else {
#>     selectColumns(x, columns, ...)
#> }
#> 
#> [[6]]
#> if (is.null(pattern)) {
#>     stop("Either 'columns' or 'pattern' must be given to removeColumns()", 
#>         call. = FALSE)
#> } else {
#>     columns <- grep(pattern, all_columns, value = TRUE)
#> }
#> 
#> [[7]]
#> if (any(isEmpty)) {
#>     catIf(dbg, sprintf("%s: %d empty columns removed: %s\n", 
#>         objectName, sum(isEmpty), paste(names(x)[isEmpty], collapse = ", ")))
#> } else {
#>     catIf(dbg, sprintf("%s: No empty columns.\n", objectName))
#> }
#> 
#> [[8]]
#> if (is.list(digits)) {
#>     digits[[column]]
#> } else {
#>     digits
#> }
#> 
#> [[9]]
#> if (is.null(x1)) {
#>     x2
#> } else {
#>     cbind(x1, x2)
#> }
#> 
#> [[10]]
#> if (is.null(pattern)) {
#>     return(x)
#> } else {
#>     columns <- grep(pattern, names(x), value = TRUE)
#> }
#>